% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_format.R
\name{fmt_format}
\alias{fmt_format}
\title{Format numbers}
\usage{
fmt_format(x, spec = NULL, locale = NULL)
}
\arguments{
\item{x}{A numeric or integer vector}

\item{spec}{A \code{\link{fmt_spec}} object, or a string coerced to a
\code{fmt_spec} object using \code{\link{as_fmt_spec}}, or a list of
arguments passed to \code{fmt_spec}.}

\item{locale}{A \code{\link{fmt_locale}} object.}
}
\value{
A character vector the same length as \code{x} of formatted numbers.
}
\description{
Format numbers
}
