% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_locales.R
\docType{data}
\name{fmt_locales}
\alias{fmt_locales}
\title{Common Format Locales}
\format{A \code{list} of \code{\link{fmt_locale}} objects.}
\source{
\url{https://github.com/d3/d3-format}
}
\usage{
fmt_locales
}
\description{
Predefined \code{\link{fmt_locale}} objects for
common locales.
}
\details{
These locales are derived from the JSON
files provided by the \href{d3-format}{https://github.com/d3/d3-format/tree/master/locale}
Javascript library.
}
\examples{
# Available locales
names(fmt_locales)
# English-US locale
fmt_locales[["en_US"]]
# French locale
fmt_locales[["fr_FR"]]
}
\keyword{datasets}
