% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision_fixed}
\alias{precision_fixed}
\alias{precision_fixed_}
\title{Suggested precision for fixed decimal formats}
\source{
The \code{d3-format} function \href{https://github.com/d3/d3-format/blob/master/src/precisionFixed.js}{precisionFixed}.
}
\usage{
precision_fixed(x)

precision_fixed_(step)
}
\arguments{
\item{x}{A numeric vector to be formatted.}

\item{step}{Minimum absolute difference between values that will be formatted.}
}
\value{
An integer vector of suggest precisions. For \code{precision_fixed},
   this is length one, for \code{precision_fixed_} it is the same length
   as \code{step}.
}
\description{
Take either a numeric vector or the maximum absolute \code{step} size
between values and and return a suggested decimal precision for fixed
point formats such as \code{"f"}.
}
\details{
\code{precision_round} Returns a suggested decimal precision for format types
that round to significant digits, such as \code{"g"} and \code{"r"}.

\code{precision_fixed} Returns a suggested SI prefix.


The suggested precision is
\deqn{
p = \max \left(0, \lfloor \log_{10}\lvert d \rvert \rfloor \right)
}{max(0, floor(log10(abs(d))))}
where \eqn{d} is the maximum absolute distance between values.
This assumes that the values to be formatted are multiples of the step
size \code{d}.
}
\examples{
# step size is 1 and the suggested precision is 0
x <- c(1, 1.5, 2)
p <- precision_fixed(x)
p
fmt(x, paste0(".", p, "f"))
# alternatively calculate the step size
step <- max(diff(sort(x)))
precision_fixed_(step)
# For 1, 2, 3, the step size is 1 suggested precision is 0
x2 <- 1:3
p2 <- precision_fixed(x2)
p2
fmt(x2, paste0(".", p, "f"))
}
\seealso{
\code{\link{precision_round}} for the suggested precision to
  use with formats that round to significant digits,
  and \code{\link{precision_prefix}} for the suggested SI prefix to use with
  formatting.
}
