% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_prefixes.R
\name{si_prefix}
\alias{si_prefix}
\alias{si_prefix.character}
\alias{si_prefix.integer}
\alias{si_prefix.numeric}
\title{Lookup SI prefix}
\usage{
si_prefix(x)

\method{si_prefix}{character}(x)

\method{si_prefix}{integer}(x)

\method{si_prefix}{numeric}(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A named character vector where values are the SI prefix exponents
  and names are the SI prefix names, from \code{\link{SI_PREFIXES}}.
}
\description{
Lookup an SI prefix by name, by exponent (power of 10), or by value.
For exponents or values, the best SI prefix is returned.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: \code{x} are the names of the prefixes. The prefix
mu can be referred to with its unicode value, \code{"\u03BC"}, or \code{"mu"}.
Empty or missing values map to no strings. Invalid values return \code{NA}.

\item \code{integer}: \code{x} are the exponents of the SI prefixes,
\eqn{-24, -21, ..., -3, 0, 3, ..., 21, 24}.

\item \code{numeric}: \code{x} are numeric values.
}}

\examples{
# lookup by name
si_prefix(c("k", "T", "mu", "\\u03BC", "", NA))
# lookup by exponent with integers
si_prefix(c(-2L, -1L, 0L, 5L))
# lookup by value with numeric vectors
si_prefix(c(-1100, 5, 1.5e8))
}
