% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_spec.R
\name{fmt_locale}
\alias{fmt_locale}
\title{Format locale}
\usage{
fmt_locale(decimal_mark = ".", grouping_mark = ",", grouping = 3,
  currency = c("", ""), numerals = NULL)
}
\arguments{
\item{decimal_mark}{string. the decimal point mark (e.g., `"."``).}

\item{grouping_mark}{string. the grouping mark (e.g., `","``).}

\item{grouping}{numeric vector group sizes (e.g., c(3L)). It is recycled as needed.}

\item{currency}{character vector of length two with the currency prefix and suffix (e.g., `c("$", "")`}

\item{numerals}{A character vector of length ten to replace the numerals 0-9.}
}
\value{
An object of class `"fmt_locale"`, which is a named list with elements: `decimal`, `thousands`, `grouping`, `currency`, and `numerals` (optional).
}
\description{
Create a a format locale. This object provides information on the symbols for decimal points, group (thousands) seperator, group sizes, currency symbols, and numerals.
}
