% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_format.R
\name{fmt_new}
\alias{fmt_new}
\alias{fmt}
\title{Format numbers}
\usage{
fmt_new(spec = NULL, locale = NULL, si = NULL)

fmt(x, spec = NULL, locale = NULL, si = NULL)
}
\arguments{
\item{spec}{A \code{\link{fmt_spec}} object, or a string coerced to a
\code{fmt_spec} object using \code{\link{as_fmt_spec}}, or a list of
arguments passed to \code{fmt_spec}.}

\item{locale}{A \code{\link{fmt_locale}} object.}

\item{si}{\code{NULL} or a numeric vector with length one.
The \href{SI-prefix}{https://en.wikipedia.org/wiki/Metric_prefix#List_of_SI_prefixes}
to use \code{x}. Unlike the \code{"s"} format type, this applies the same
prefix to all values in \code{x}, rather than determining the SI-prefix
on a value by value basis. Additionally, a non-\code{NULL} \code{si} implies
a \code{"f"} format type, in which \code{precision} represents the number of
digits past the decimal point.}

\item{x}{A numeric or integer vector}
}
\value{
\code{fmt_new} returns a function with a single argument.
   \code{fmt} returns a function of the same length as \code{x} of
   formatted numbers.
}
\description{
The function \code{fmt_new} creates a function to format numbers using
a given \code{locale} and \code{spec}. The function \code{fmt} formats
numbers, and is a convenience function for \code{fmt_new(...)(x)}.
}
\details{
The supported SI-prefixes are:
\itemize{
\item{\code{"y"} - yocto, \eqn{10^{-24}}}
\item{\code{"z"} - zepto, \eqn{10^{-21}}}
\item{\code{"a"} - atto, \eqn{10^{-18}}}
\item{\code{"f"} - femto, \eqn{10^{-15}}}
\item{\code{"p"} - pico, \eqn{10^{-12}}}
\item{\code{"n"} - nano, \eqn{10^{-9}}}
\item{\code{"µ"} - micro, \eqn{10^{-6}}}
\item{\code{"m"} - milli, \eqn{10^{-3}}}
\item{\code{" "} (none) - \eqn{10^0}}
\item{\code{"k"} - kilo, \eqn{10^3}}
\item{\code{"M"} - mega, \eqn{10^{6}}}
\item{\code{"G"} - giga, \eqn{10^{9}}}
\item{\code{"T"} - tera, \eqn{10^{12}}}
\item{\code{"P"} - peta, \eqn{10^{15}}}
\item{\code{"E"} - exa, \eqn{10^{18}}}
\item{\code{"Z"} - zetta, \eqn{10^{21}}}
\item{\code{"Y"} - yotta, \eqn{10^{24}}}
}
}
\examples{
fmt(c(0.00042, 0.0042), spec = ",.0", si = 1e-6)
}
