% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_new}
\alias{fmt_new}
\alias{fmt}
\title{Format numbers}
\usage{
fmt_new(spec = NULL, locale = NULL, si_prefix = NULL)

fmt(x, spec = NULL, locale = NULL, ...)
}
\arguments{
\item{spec}{A \code{\link{fmt_spec}} object, or a string coerced to a
\code{fmt_spec} object using \code{\link{as_fmt_spec}}, or a list of
arguments passed to \code{fmt_spec}.}

\item{locale}{A \code{\link{fmt_locale}} object.}

\item{si_prefix}{If non-\code{NULL}, then use
an \href{https://en.wikipedia.org/wiki/Metric_prefix#List_of_SI_prefixes}{SI prefix} to format \code{x}.
If \code{TRUE}, then use \code{precision_prefix} to automatically determine
the SI prefix to use. Otherwise, it can be any of the valid arguments
for \code{\link{si_prefix}}: a string with the prefix name, an integer with the
SI prefix exponent, or a numeric value.
Unlike the \code{"s"} format type, this applies the same
prefix to all values in \code{x}, rather than determining the SI-prefix
on a value by value basis. Additionally, a non-\code{NULL} \code{si} implies
a \code{"f"} format type, in which \code{precision} represents the number of
digits past the decimal point.}

\item{x}{A numeric or integer vector}

\item{...}{Arguments passed to \code{fmt_new}.}
}
\value{
\code{fmt_new} returns a function with a single argument.
   \code{fmt} returns a function of the same length as \code{x} of
   formatted numbers.
}
\description{
The function \code{fmt_new} creates a function to format numbers using
a given \code{locale} and \code{spec}. The function \code{fmt} formats
numbers, and is a convenience function for \code{fmt_new(...)(x)}.
}
\examples{
fmt(c(0.00042, 0.0042), spec = ",.0", si_prefix = 1e-6)
}
