% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_spec.R
\name{as_fmt_spec}
\alias{as_fmt_spec}
\alias{fmt_spec}
\title{Format Specification}
\usage{
as_fmt_spec(x)

fmt_spec(fill = NULL, align = c(">", "<", "^", "="), sign = c("-", "+",
  "(", " "), symbol = NULL, zero = FALSE, width = NA_integer_,
  comma = FALSE, precision = NA_integer_, type = NULL)
}
\arguments{
\item{x}{A string representation of the spec. See Details.}

\item{fill}{A string. See Details.}

\item{align}{A string. See Details.}

\item{sign}{A string. See Details.}

\item{symbol}{A string. See Details.}

\item{zero}{A logical vector of length one. See Details.}

\item{width}{An integer vector of length one. See Details.}

\item{comma}{An logical vector of length one. See Details.}

\item{precision}{An integer vector of length one. See Details.}

\item{type}{A character vector of length one. See Details.}
}
\value{
An object of class \code{"fmt_spec"}. This is a list with elements:
\describe{
\item{fill}{Character vector of length one. See Details.}
\item{align}{Character vector of length one. See Details.}
\item{sign}{Character vector of length one. See Details.}
\item{symbol}{Character vector of length one. See Details.}
\item{width}{Integer vector of length one. See Details.}
\item{comma}{Logical vector of length one. See Details.}
\item{precision}{Integer vector of length one.}
\item{type}{Character vector of length one, or \code{NULL} if the
            default type}
}
}
\description{
Format Specification
}
