% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_spec.R
\name{as_fmt_spec}
\alias{as_fmt_spec}
\alias{fmt_spec}
\title{Format Spec}
\usage{
as_fmt_spec(x)

fmt_spec(fill = NULL, align = c(">", "<", "^", "="), sign = c("-", "+",
  "(", " "), symbol = NULL, zero = FALSE, width = NA_integer_,
  comma = FALSE, precision = NA_integer_, type = NULL)
}
\arguments{
\item{x}{An object to convert to a format spec: a string representation
of the spec.}

\item{fill}{A string. See details.}

\item{align}{A string. See Details.}

\item{sign}{A string. See Details.}

\item{symbol}{A string. See Details.}
}
\description{
Format Spec
}
