% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision_fixed}
\alias{precision_fixed}
\alias{precision_fixed_}
\alias{precision_round}
\alias{precision_round_}
\alias{precision_prefix}
\alias{precision_prefix_}
\title{Suggested precisions}
\usage{
precision_fixed(x)

precision_fixed_(step)

precision_round(x)

precision_round_(step, .max)

precision_prefix(x, value)

precision_prefix_(step, value)
}
\arguments{
\item{x}{A numeric vector}

\item{step}{Minimum absolute difference between values that will be formatted}

\item{.max}{Largest absolute value to be formatted}

\item{value}{The SI prefix to be used}
}
\value{
An integer vector of decimal precisions
}
\description{
\code{precision_prefix} returns a suggested decimal precision for fixed point
notation.
}
\details{
\code{precision_round} Returns a suggested decimal precision for format types
that round to significant digits given the specified numeric \code{step} and
\code{max} values.

\code{precision_fixed} returns a suggested decimal precision for a
format precision.
}
