% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision_prefix}
\alias{precision_prefix}
\alias{precision_prefix_}
\title{Suggested precision for SI Prefix formatting}
\source{
The \code{d3-format} function \href{https://github.com/d3/d3-format/blob/master/src/precisionPrefix.js}{precisionPrefix}.
}
\usage{
precision_prefix(x, prefix = NULL)

precision_prefix_(step, prefix)
}
\arguments{
\item{x}{A numeric vector of values to be formatted.}

\item{prefix}{The SI prefix to use. This is any valid argument for
\code{\link{si_prefix}}. If \code{NULL} in \code{precision_prefix},
then the prefix is determined from the median value in \code{x}.}

\item{step}{Numeric: Minimum absolute difference between values that will be formatted.}
}
\value{
Named list of two elements:
\describe{
\item{precision}{Integer vector: The suggested precision}
\item{si_prefix}{An named integer vector of SI prefixes. The names are the
  SI prefix names, and the values are the exponents, as in
  \code{\link{SI_PREFIXES}}.}
}
}
\description{
The function \code{precision_prefix_} provides a suggested precision for
use with SI prefix to use,
given a \code{value} to format, and the difference between the values,
\code{step}, that will be formatted.
\code{precision_prefix} calculates the suggested SI Prefix given a
numeric vector of values to format, \code{x}.
}
\details{
The suggested precision, \eqn{p}, when formatting a vector with
an SI Prefix of \eqn{10^k} is,
\deqn{
p = \max \left( 0, k - \lfloor \log_{10}\lvert d \rvert \rfloor \right)
,}{p = max(0, floor(log10(d))),}
where \eqn{d} is the maximum absolute value between elements in the vector.
}
\examples{
x <- c(1.1e6, 1.2e6, 1.3e6)
p <- precision_prefix(x, 6L)
p$precision
fmt(x, paste0(".", p$precision), si_prefix = 6L)
}
\seealso{
\code{\link{precision_fixed}} for the suggested precision to
  use with formats that use a fixed number of digits after the decimal point,
  and \code{\link{precision_round}} for the suggested precision to use with
  formats that round to significant digits.
}
