% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision_round}
\alias{precision_round}
\alias{precision_round_}
\title{Suggested precision for rounded significant digit formats}
\source{
The \code{d3-format} function \href{https://github.com/d3/d3-format/blob/master/src/precisionRound.js}{precisionRound}
}
\usage{
precision_round(x)

precision_round_(step, xmax)
}
\arguments{
\item{x}{A numeric vector of values to be formatted.}

\item{step}{Numeric: Minimum absolute difference between values that will be formatted.}

\item{xmax}{Numeric: Maximum absolute value of the values to be formatted.}
}
\value{
An integer vector of suggest precisions. For \code{precision_fixed},
   this is length one, for \code{precision_fixed_} it is the same length
   as \code{step}.
}
\description{
The function \code{precision_round_} takes the maximum absolute distance
between values, (\code{step}), and the maximum absolute value, (\code{value}),
of the values to be formatted, and returns the suggested precision for
formats that round to significant digits (e.g. \code{"g"}).
\code{precision_round} calculates this precision given a numeric vector of
values to format, \code{x}.
}
\seealso{
\code{\link{precision_fixed}} for the suggested precision to
  use with formats that use a fixed number of digits after the decimal point,
  and \code{\link{precision_prefix}} for a suggested SI prefix to use.
}
