% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_format.R
\docType{data}
\name{SI_PREFIXES}
\alias{SI_PREFIXES}
\title{SI Prefixes}
\format{A named numeric vector of SI prefixes.
The names are the SI prefixes, the values are the exponents,
\eqn{10^{-24}, 10^{-21}, \dots, 0, \dots, 10^{24}}{10^24, 10^21, ..., 0, ..., 10^24}.}
\source{
\href{https://en.wikipedia.org/wiki/Metric_prefix\#List_of_SI_prefixes}{Table of SI prefixes}
}
\usage{
SI_PREFIXES
}
\description{
SI Prefixes
}
\examples{
SI_PREFIXES
}
\keyword{datasets}
